Create a shared `ReplayError` type
Replay errors for both send and receive are almost identical. This
commit creates a shared error abstraction with two generic params
(SessionEvent, SessionState).
